refactor: update zone and source models for improved structure and clarity

+++ apps/frontend/flutter-apps/fusion_app/lib/core/models/gain_model.dart
@@ -22,7 +22,7 @@ class GainConfig {
   }
 }
 class GainValue {
-  final int gain;
+  final double gain;
   final bool mute;
 
   GainValue({
